pyproject.toml
[project]
name = "belief-graph-persuasion"
version = "0.1.0"
description = "Multi-agent system for modeling political beliefs and generating nudge-based persuasion strategies"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.2.0",
    "networkx>=3.2",
    "numpy>=1.26",
    "pydantic>=2.5",
    "python-dotenv>=1.0",
    "structlog>=24.0",
    "typer>=0.12.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "ruff>=0.3.0",
    "mypy>=1.8",
]

[project.scripts]
persuasion = "scripts.run_simulation:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]
